pyproject.toml
[project]
name = "ilink-digital"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.34.0",
    "gtts>=2.5.4",
    "jupyter>=1.1.1",
    "langgraph>=0.6.7",
    "numpy>=2.2.6",
    "openai-whisper>=20250625",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.49.1",
    "torch>=2.8.0",
]
[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
